+++ src/main/java/com/atwoz/mission/ui/membermission/dto/RewardResponse.java
@@ -1,4 +1,4 @@
-package com.atwoz.mission.ui.membermissions.dto;
+package com.atwoz.mission.ui.membermission.dto;
 
 public record RewardResponse(
         Integer reward
